langchain_sumpy_openai/requirements.txt
#openai
#langchain
sympy
tiktoken
langchain-community
langchain-openai